Crate double_int

The double-int format represents an integer that can be stored in an IEEE 754 double-precision number without loss of precision.

This crate has been designed for use with OpenAPI tooling that wish to support integer-based format: double-int fields. See docs in the OpenAPI format registry.

Examples

#[derive(Debug, serde::Deserialize)]
struct Config {
    count: DoubleInt,
}

let config = toml::from_str::<Config>(r#"
    count = 42
"#).unwrap();
assert_eq!(config.count, 42);

let config = toml::from_str::<Config>(r#"
    count = -42
"#).unwrap();
assert_eq!(config.count, -42);

// count is outside the bounds of a double-int (> 2^53 in this case)
// (this would usually be accepted by an i64)
let config = toml::from_str::<Config>(r#"
    count = 36028797018963968
"#).unwrap_err();

Structs

DoubleInt

Type that only deserializes from the true boolean value.